requirements.txt
streamlit
openai
sentence-transformers
pinecone-client
pandas
matplotlib
seaborn
plotly
python-dotenv
scikit-learn
yfinance
nltk
scipy
